perf(home): rename Blog as LatestNodes and BlogPost as NoteCard
