+++ src/home/LatestNotes/NoteCard/styles.ts
